* Initial commit

* feat(api): api update

* feat(api): api update

* feat(api): api update

* feat(api): api update

* feat(api): api update

* feat(api): api update

* feat(api): api update

* feat(api): manual updates

* feat(api): api update

* feat(api): manual updates

* chore: rebuild project due to codegen change

* chore: rebuild project due to codegen change

* feat(api): api update

* chore: rebuild project due to codegen change

* chore: rebuild project due to codegen change

* codegen metadata

* chore(internal): fix compat model_dump method when warnings are passed

* docs: add info log level to readme

* chore: remove now unused `cached-property` dep

* chore(internal): exclude mypy from running on tests

* fix(client): compat with new httpx 0.28.0 release

* feat(api): api update

* chore(internal): bump pyright

* chore: make the `Omit` type public

* feat(api): api update

* feat(api): api update

* chore(internal): bump pydantic dependency

* docs(readme): fix http client proxies example

* feat(api): api update

* feat(api): api update

* chore(internal): bump pyright

* chore(internal): add support for TypeAliasType

* feat(api): api update

* chore(internal): codegen related update

* chore(internal): codegen related update

* chore(internal): codegen related update

* docs(readme): example snippet for client context manager

* chore(internal): fix some typos

* codegen metadata

* feat(api): api update

* chore(internal): codegen related update

* chore(internal): bump httpx dependency

* fix(client): only call .close() when needed

* docs: fix typos

* chore(internal): codegen related update

* chore(internal): codegen related update

* feat(api): api update

* chore(internal): codegen related update

* chore(internal): codegen related update

* feat(api): api update

* chore(internal): codegen related update

* chore(internal): minor formatting changes

* chore: update test examples

* chore(internal): change default timeout to an int

* chore(internal): bummp ruff dependency

* feat(api): api update

* feat(api): api update

* feat(client): send `X-Stainless-Read-Timeout` header

* chore(internal): fix type traversing dictionary params

* chore(internal): minor type handling changes

* chore(internal): update client tests

* fix: asyncify on non-asyncio runtimes

* feat(client): allow passing `NotGiven` for body
fix(client): mark some request bodies as optional

* chore(internal): fix devcontainers setup

* chore(internal): properly set __pydantic_private__

* docs: update URLs from stainlessapi.com to stainless.com

More details at https://www.stainless.com/changelog/stainless-com

* chore(docs): update client docstring

* chore(internal): remove unused http client options forwarding

* feat(api): api update

* chore(internal): codegen related update

* chore(internal): slight transform perf improvement

* chore(internal): expand CI branch coverage

* chore(internal): reduce CI branch coverage

* fix(perf): skip traversing types for NotGiven values

* fix(perf): optimize some hot paths

* feat(api): api update

---------

+++ tests/test_client.py
@@ -6,6 +6,7 @@ import gc
 import os
 import sys
 import json
+import time
 import asyncio
 import inspect
 import subprocess
@@ -22,6 +23,7 @@ from pydantic import ValidationError
 
 from agility import Agility, AsyncAgility, APIResponseValidationError
 from agility._types import Omit
+from agility._utils import maybe_transform
 from agility._models import BaseModel, FinalRequestOptions
 from agility._constants import RAW_RESPONSE_HEADER
 from agility._exceptions import APIStatusError, APITimeoutError, APIResponseValidationError
@@ -31,6 +33,7 @@ from agility._base_client import (
     BaseClient,
     make_request_options,
 )
+from agility.types.assistant_create_params import AssistantCreateParams
 
 from .utils import update_env
 
@@ -714,8 +717,13 @@ class TestAgility:
                 "/api/assistants/",
                 body=cast(
                     object,
-                    dict(
-                        description="description", knowledge_base_id="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e", name="name"
+                    maybe_transform(
+                        dict(
+                            description="description",
+                            knowledge_base_id="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
+                            name="name",
+                        ),
+                        AssistantCreateParams,
                     ),
                 ),
                 cast_to=httpx.Response,
@@ -734,8 +742,13 @@ class TestAgility:
                 "/api/assistants/",
                 body=cast(
                     object,
-                    dict(
-                        description="description", knowledge_base_id="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e", name="name"
+                    maybe_transform(
+                        dict(
+                            description="description",
+                            knowledge_base_id="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
+                            name="name",
+                        ),
+                        AssistantCreateParams,
                     ),
                 ),
                 cast_to=httpx.Response,
@@ -1512,8 +1525,13 @@ class TestAsyncAgility:
                 "/api/assistants/",
                 body=cast(
                     object,
-                    dict(
-                        description="description", knowledge_base_id="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e", name="name"
+                    maybe_transform(
+                        dict(
+                            description="description",
+                            knowledge_base_id="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
+                            name="name",
+                        ),
+                        AssistantCreateParams,
                     ),
                 ),
                 cast_to=httpx.Response,
@@ -1532,8 +1550,13 @@ class TestAsyncAgility:
                 "/api/assistants/",
                 body=cast(
                     object,
-                    dict(
-                        description="description", knowledge_base_id="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e", name="name"
+                    maybe_transform(
+                        dict(
+                            description="description",
+                            knowledge_base_id="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
+                            name="name",
+                        ),
+                        AssistantCreateParams,
                     ),
                 ),
                 cast_to=httpx.Response,
@@ -1646,7 +1669,7 @@ class TestAsyncAgility:
         import threading
 
         from agility._utils import asyncify
-        from agility._base_client import get_platform 
+        from agility._base_client import get_platform
 
         async def test_main() -> None:
             result = await asyncify(get_platform)()
@@ -1661,10 +1684,20 @@ class TestAsyncAgility:
             [sys.executable, "-c", test_code],
             text=True,
         ) as process:
-            try:
-                process.wait(2)
-                if process.returncode:
-                    raise AssertionError("calling get_platform using asyncify resulted in a non-zero exit code")
-            except subprocess.TimeoutExpired as e:
-                process.kill()
-                raise AssertionError("calling get_platform using asyncify resulted in a hung process") from e
+            timeout = 10  # seconds
+
+            start_time = time.monotonic()
+            while True:
+                return_code = process.poll()
+                if return_code is not None:
+                    if return_code != 0:
+                        raise AssertionError("calling get_platform using asyncify resulted in a non-zero exit code")
+
+                    # success
+                    break
+
+                if time.monotonic() - start_time > timeout:
+                    process.kill()
+                    raise AssertionError("calling get_platform using asyncify resulted in a hung process")
+
+                time.sleep(0.1)
